Move comit_wallet and ethereum_wallet into comit_node
