+++ Cargo.toml
@@ -3,8 +3,6 @@
 members = [
     "application/comit_node",
     "application/comit_node_client",
-    "application/comit_wallet",
-    "application/ethereum_wallet",
     "application/ledger_query_service",
     "vendor/bitcoin_rpc_test_helpers",
     "vendor/bitcoin_support",
