+++ application/comit_node/src/bin/comit_node.rs
@@ -4,9 +4,7 @@
 extern crate bitcoin_rpc_client;
 extern crate bitcoin_support;
 extern crate comit_node;
-extern crate comit_wallet;
 extern crate ethereum_support;
-extern crate ethereum_wallet;
 #[macro_use]
 extern crate log;
 extern crate event_store;
@@ -20,8 +18,10 @@ use comit_node::{
     bitcoin_fee_service::StaticBitcoinFeeService,
     comit_client,
     comit_server::ComitServer,
+    ethereum_wallet::InMemoryWallet,
     gas_price_service::StaticGasPriceService,
     gotham_factory,
+    key_store::KeyStore,
     ledger_query_service::DefaultLedgerQueryServiceApiClient,
     logging,
     settings::ComitNodeSettings,
@@ -30,9 +30,7 @@ use comit_node::{
         ledger_htlc_service::{BitcoinService, EthereumService},
     },
 };
-use comit_wallet::KeyStore;
 use ethereum_support::*;
-use ethereum_wallet::InMemoryWallet;
 use event_store::InMemoryEventStore;
 use std::{env::var, net::SocketAddr, str::FromStr, sync::Arc};
 use web3::{transports::Http, Web3};
