+++ application/comit_node/src/ethereum_wallet/mod.rs
@@ -1,0 +1,9 @@
+mod transaction;
+mod wallet;
+
+pub use self::{
+    transaction::*,
+    wallet::{InMemoryWallet, Wallet},
+};
+
+pub mod fake;
