+++ application/comit_node/src/key_store/mod.rs
@@ -1,0 +1,7 @@
+mod fake_key_store;
+mod key_store;
+
+pub use self::{
+    fake_key_store::FakeKeyStoreFactory,
+    key_store::{Error, KeyStore},
+};
