+++ application/comit_node/src/lib.rs
@@ -25,11 +25,9 @@ extern crate bitcoin_rpc_client;
 extern crate bitcoin_support;
 extern crate bitcoin_witness;
 extern crate chrono;
-extern crate comit_wallet;
 extern crate config;
 extern crate crypto;
 extern crate ethereum_support;
-extern crate ethereum_wallet;
 extern crate event_store;
 extern crate fern;
 extern crate futures;
@@ -40,10 +38,12 @@ extern crate hyper;
 extern crate mime;
 extern crate rand;
 extern crate reqwest;
+extern crate rlp;
 extern crate rustc_hex;
 extern crate secp256k1_support;
 extern crate serde;
 extern crate serde_json;
+extern crate tiny_keccak;
 extern crate tokio;
 extern crate tokio_timer;
 extern crate url;
@@ -53,9 +53,11 @@ pub mod bitcoin_fee_service;
 
 pub mod comit_client;
 pub mod comit_server;
+pub mod ethereum_wallet;
 pub mod gas_price_service;
 pub mod gotham_factory;
 pub mod http_api;
+pub mod key_store;
 pub mod ledger_query_service;
 pub mod logging;
 pub mod settings;
