+++ application/comit_node/tests/parity_swap_tests.rs
@@ -3,7 +3,6 @@
 
 extern crate comit_node;
 extern crate ethereum_support;
-extern crate ethereum_wallet;
 extern crate hex;
 extern crate pretty_env_logger;
 extern crate secp256k1_support;
