--- application/ethereum_wallet/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "ethereum_wallet"
-version = "0.1.0"
-authors = [
-    "Thomas Eizinger <thomas@coblox.tech>",
-    "Philipp Hoenisch <philipp@coblox.tech>",
-    "Franck Royer <franck@coblox.tech>",
-    "Lloyd Fournier <lloyd@coblox.tech>",
-]
-
-[dependencies]
-ethereum_support = { path = "../../vendor/ethereum_support" }
-secp256k1_support = { path = "../../vendor/secp256k1_support" }
-hex = "0.3.2"
-rlp = "0.2.4"
-tiny-keccak = "1.4.2"
-
-[dev-dependencies]
-hex = "0.3.1"
-testcontainers = "0.3"
-tc_web3_client = { path = "../../vendor/tc_web3_client" }
-env_logger = "0.5.11"
