--- application/ethereum_wallet/src/lib.rs
@@ -1,16 +1,0 @@
-#![warn(unused_extern_crates, missing_debug_implementations)]
-#![deny(unsafe_code)]
-
-extern crate ethereum_support;
-extern crate hex;
-extern crate rlp;
-extern crate secp256k1_support;
-extern crate tiny_keccak;
-
-mod transaction;
-mod wallet;
-
-pub use transaction::*;
-pub use wallet::{InMemoryWallet, Wallet};
-
-pub mod fake;
